+++ autode/conformers/conf_gen.py
@@ -41,8 +41,8 @@ def get_bond_matrix(n_atoms, bonds, fixed_bonds):
     return bond_matrix
 
 
-def get_coords_minimised_v(coords, bonds, k, c, d0, tol, fixed_bonds,
-                           exponent=8, return_energy=False):
+def get_coords_energy(coords, bonds, k, c, d0, tol, fixed_bonds,
+                      exponent=8):
     """
     Get the coordinates that minimise a FF with a bonds + repulsion FF
     where the repulsion is c/r^exponent
@@ -57,7 +57,6 @@ def get_coords_minimised_v(coords, bonds, k, c, d0, tol, fixed_bonds,
 
     Keyword Arguments:
         exponent (int): Exponent in the repulsive pairwise term
-        return_energy (bool): Whether to also return the energy
 
     Returns:
         (np.ndarray): Optimised coordinates, shape = (n_atoms, 3)
@@ -69,22 +68,18 @@ def get_coords_minimised_v(coords, bonds, k, c, d0, tol, fixed_bonds,
     n_atoms = len(coords)
     os.environ['OMP_NUM_THREADS'] = str(1)
 
-    init_coords = coords.reshape(3 * n_atoms)
     bond_matrix = get_bond_matrix(n_atoms=len(coords),
                                   bonds=bonds,
                                   fixed_bonds=fixed_bonds)
 
-    res = minimize(v, x0=init_coords,
+    res = minimize(v,
+                   x0=coords.reshape(3 * n_atoms),
                    args=(bond_matrix, k, d0, c, exponent),
                    method='CG',
                    tol=tol,
                    jac=dvdr)
-    coords = res.x.reshape(n_atoms, 3)
 
-    if return_energy:
-        return coords, res.fun
-
-    return coords
+    return res.x.reshape(n_atoms, 3), res.fun
 
 
 def get_v(coords, bonds, k, c, d0, fixed_bonds, exponent=8):
@@ -296,28 +291,29 @@ def get_coords_no_init_structure(atoms, species, d0, constrained_bonds):
         (np.ndarray): Optimised coordinates, shape = (n_atoms, 3)
     """
     # Minimise atoms with no bonds between them
-    far_coords = get_coords_minimised_v(coords=np.array([atom.coord for atom in atoms]),
-                                        bonds=species.graph.edges,
-                                        fixed_bonds=constrained_bonds,
-                                        k=0.0, c=0.1, d0=d0, tol=5E-3,
-                                        exponent=2)
+    far_coords, _ = get_coords_energy(coords=np.array([atom.coord
+                                                       for atom in atoms]),
+                                      bonds=species.graph.edges,
+                                      fixed_bonds=constrained_bonds,
+                                      k=0.0, c=0.1, d0=d0, tol=5E-3,
+                                      exponent=2)
     coords = far_coords[:2]
 
     # Add the atoms one by one to the structure. Thanks to Dr. Cyrille Lavigne
     #  for this suggestion!
     for n in range(2, species.n_atoms):
-        coords = get_coords_minimised_v(np.concatenate((coords, far_coords[len(coords):n+1])),
-                                        bonds=species.graph.edges,
-                                        fixed_bonds=constrained_bonds,
-                                        k=0.1, c=0.1, d0=d0, tol=1E-3,
-                                        exponent=2)
+        new_coords = np.concatenate((coords, far_coords[len(coords):n + 1]))
+        coords, _ = get_coords_energy(new_coords,
+                                      bonds=species.graph.edges,
+                                      fixed_bonds=constrained_bonds,
+                                      k=0.1, c=0.1, d0=d0, tol=1E-3,
+                                      exponent=2)
 
     # Perform a final minimisation
-    coords, energy = get_coords_minimised_v(coords=coords,
-                                            bonds=species.graph.edges,
-                                            fixed_bonds=constrained_bonds,
-                                            k=1.0, c=0.01, d0=d0, tol=1E-5,
-                                            return_energy=True)
+    coords, energy = get_coords_energy(coords=coords,
+                                       bonds=species.graph.edges,
+                                       fixed_bonds=constrained_bonds,
+                                       k=1.0, c=0.01, d0=d0, tol=1E-5)
     return coords, energy
 
 
@@ -405,11 +401,10 @@ def get_simanl_atoms(species, dist_consts=None, conf_n=0, save_xyz=True,
     st = time()
     if reasonable_init_coords:
         init_coords = np.array([atom.coord for atom in atoms])
-        coords, energy = get_coords_minimised_v(coords=init_coords,
-                                                bonds=species.graph.edges,
-                                                k=1.0, c=0.01, d0=d0, tol=1E-5,
-                                                fixed_bonds=constrained_bonds,
-                                                return_energy=True)
+        coords, energy = get_coords_energy(coords=init_coords,
+                                           bonds=species.graph.edges,
+                                           k=1.0, c=0.01, d0=d0, tol=1E-5,
+                                           fixed_bonds=constrained_bonds)
     else:
         coords, energy = get_coords_no_init_structure(atoms, species, d0,
                                                       constrained_bonds)
